Add conversation_start
