+++ conversation.go
@@ -1,10 +1,6 @@
 package client
 
 import (
-	"context"
-	"encoding/json"
-	"fmt"
-	"net/http"
 	"time"
 )
 
@@ -48,29 +44,6 @@ const (
 	StatusFailed Status = "failed"
 )
 
-// StartConversationParams are the parameters to Client.StartConversation.
-type StartConversationParams struct {
-	// ID uniquely identifies the conversation.
-	//
-	// Can be anything consisting of letters, numbers, or any of the following
-	// characters: _ - + =.
-	//
-	// Tip: use something meaningful to your business (e.g. a ticket number).
-	ID string `json:"id"`
-
-	// CustomerID uniquely identifies the customer. Used to build historical
-	// context of conversations the agent has had with this customer.
-	CustomerID string `json:"customer_id"`
-
-	// Channel represents the way a customer is getting in touch. It will be used
-	// to determine how the agent formats responses, etc.
-	Channel Channel `json:"channel"`
-
-	// Metadata is arbitrary metadata that will be attached to the conversation.
-	// It will be passed along with webhooks so can be used as action parameters.
-	Metadata any `json:"metadata"`
-}
-
 // Conversation represents a series of messages between a customer, human agent,
 // and the AI Agent.
 type Conversation struct {
@@ -102,39 +75,4 @@ type Conversation struct {
 
 	// Status describes the current state of the conversation.
 	Status Status `json:"status"`
-}
-
-// StartConversation begins a conversation with the agent.
-func (c *Client) StartConversation(ctx context.Context, p StartConversationParams) (*Conversation, error) {
-	rsp, err := c.makeRequest(ctx, http.MethodPost, "conversations", p)
-	if err != nil {
-		return nil, err
-	}
-	defer rsp.Body.Close()
-
-	if rsp.StatusCode < 200 || rsp.StatusCode > 299 {
-		return nil, newResponseError(rsp)
-	}
-
-	var conv Conversation
-	if err := json.NewDecoder(rsp.Body).Decode(&conv); err != nil {
-		return nil, err
-	}
-	return &conv, nil
-}
-
-// CancelConversation cancels a conversation to abruptly stop the agent
-// responding (e.g. because a human agent has taken over).
-func (c *Client) CancelConversation(ctx context.Context, conversationID string) error {
-	rsp, err := c.makeRequest(ctx, http.MethodPut, fmt.Sprintf("conversations/%s/cancel", conversationID), nil)
-	if err != nil {
-		return err
-	}
-	defer rsp.Body.Close()
-
-	if rsp.StatusCode < 200 || rsp.StatusCode > 299 {
-		return newResponseError(rsp)
-	}
-
-	return nil
 }
